chore: clean unused code/assets; refactor dark mode context/hook; fix imports
