--- old/CreateCabinForm-old.jsx
@@ -1,128 +1,0 @@
-import Input from "../../ui/Input";
-import Form from "../../ui/Form";
-import Button from "../../ui/Button";
-import FileInput from "../../ui/FileInput";
-import Textarea from "../../ui/Textarea";
-import FormRow from "../../ui/FormRow";
-
-import { useCabinForm } from "./useCabinForm";
-import { useCreateCabin } from "./useCreateCabin";
-import { useEditCabin } from "./useEditCabin";
-
-function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
-  const { createCabin } = useCreateCabin();
-  const { editCabin } = useEditCabin();
-
-  const { register, getValues, isWorking, onSubmit, errors } = useCabinForm({
-    cabinToEdit,
-    onSuccess: onCloseModal,
-    createCabin,
-    editCabin,
-  });
-
-  const isEditSession = Boolean(cabinToEdit?.id);
-
-  return (
-    <Form onSubmit={onSubmit} type={onCloseModal ? "modal" : "regular"}>
-      <FormRow label="Cabin name" error={errors?.name?.message}>
-        <Input
-          type="text"
-          id="name"
-          disabled={isWorking}
-          {...register("name", { required: "This field is required" })}
-        />
-      </FormRow>
-
-      <FormRow label="Maximum capacity" error={errors?.maxCapacity?.message}>
-        <Input
-          type="number"
-          id="maxCapacity"
-          disabled={isWorking}
-          {...register("maxCapacity", {
-            required: "This field is required",
-            min: {
-              value: 1,
-              message: "Capacity must be at least 1",
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow label="Regular price" error={errors?.regularPrice?.message}>
-        <Input
-          type="number"
-          id="regularPrice"
-          disabled={isWorking}
-          {...register("regularPrice", {
-            required: "This field is required",
-            min: {
-              value: 1,
-              message: "Regular price must be at least 1",
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow label="Discount" error={errors?.discount?.message}>
-        <Input
-          type="number"
-          id="discount"
-          disabled={isWorking}
-          {...register("discount", {
-            required: "This field is required",
-            validate: (value) => {
-              const regularPrice = parseFloat(getValues().regularPrice);
-              const discount = parseFloat(value);
-
-              if (isNaN(regularPrice))
-                return "Please enter valid regular price first";
-              if (discount < 0) return "Discount cannot be negative";
-              if (discount >= regularPrice)
-                return `Discount must be less than ${regularPrice.toFixed(2)}`;
-              return true;
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow
-        label="Description for website"
-        disabled={isWorking}
-        error={errors?.description?.message}
-      >
-        <Textarea
-          id="description"
-          disabled={isWorking}
-          {...register("description", {
-            required: "This field is required",
-          })}
-        />
-      </FormRow>
-
-      <FormRow label="Cabin photo" error={errors.image?.message}>
-        <FileInput
-          id="image"
-          accept="image/*"
-          {...register("image", {
-            required: isEditSession ? false : "This field is required",
-          })}
-        />
-      </FormRow>
-
-      <FormRow>
-        <Button
-          variation="secondary"
-          type="reset"
-          onClick={() => onCloseModal?.()}
-        >
-          Cancel
-        </Button>
-        <Button disabled={isWorking}>
-          {isEditSession ? "Edit cabin" : "Create new cabin"}
-        </Button>
-      </FormRow>
-    </Form>
-  );
-}
-
-export default CreateCabinForm;
