+++ src/App.jsx
@@ -16,7 +16,7 @@ import AppLayout from "./ui/AppLayout";
 import Booking from "./pages/Booking";
 import Checkin from "./pages/Checkin";
 import ProtectedRoute from "./ui/ProtectedRoute";
-import { DarkModeProvider } from "./context/DarkModeContext";
+import DarkModeProvider from "./context/DarkModeContext.jsx";
 
 const queryClient = new QueryClient({
   defaultOptions: {
