+++ src/context/DarkModeContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const DarkModeContext = createContext();
