+++ src/context/DarkModeContext.jsx
@@ -1,9 +1,8 @@
-import { createContext, useContext, useEffect } from "react";
+import { useEffect } from "react";
 import { useLocalStorageState } from "../hooks/useLocalStorageState";
+import { DarkModeContext } from "./DarkModeContext";
 
-const DarkModeContext = createContext();
-
-function DarkModeProvider({ children }) {
+export default function DarkModeProvider({ children }) {
   const [isDarkMode, setIsDarkMode] = useLocalStorageState(
     window.matchMedia("(prefers-color-scheme: dark)").matches,
     "isDarkMode"
@@ -32,13 +31,3 @@ function DarkModeProvider({ children }) {
     </DarkModeContext.Provider>
   );
 }
-
-function useDarkMode() {
-  const context = useContext(DarkModeContext);
-  if (context === undefined) {
-    throw new Error("DarkModeContext was used of DarkModeProvider");
-  }
-  return context;
-}
-
-export { DarkModeProvider, useDarkMode };
