--- src/features/cabins/CreateCabinForm-v1.jsx
@@ -1,137 +1,0 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import toast from "react-hot-toast";
-
-import Input from "../../ui/Input";
-import Form from "../../ui/Form";
-import Button from "../../ui/Button";
-import FileInput from "../../ui/FileInput";
-import Textarea from "../../ui/Textarea";
-import FormRow from "../../ui/FormRow";
-
-import { useForm } from "react-hook-form";
-import { createCabin } from "../../services/apiCabins";
-
-function CreateCabinForm() {
-  const { register, handleSubmit, reset, getValues, formState } = useForm();
-  const { errors } = formState;
-  const queryClient = useQueryClient();
-
-  const { mutate, isLoading: isCreating } = useMutation({
-    mutationFn: createCabin,
-    onSuccess: () => {
-      toast.success("Cabin created successfully");
-      queryClient.invalidateQueries({
-        queryKey: ["cabins"],
-      });
-      reset();
-    },
-    onError: (error) => {
-      toast.error("Cabin could not be created");
-      console.error(error);
-    },
-  });
-
-  function onSubmit(data) {
-    mutate({ ...data, image: data.image[0] });
-  }
-
-  function onError(errors) {
-    console.log(errors);
-  }
-
-  return (
-    <Form onSubmit={handleSubmit(onSubmit, onError)}>
-      <FormRow label="Cabin name" error={errors?.name?.message}>
-        <Input
-          type="text"
-          id="name"
-          disabled={isCreating}
-          {...register("name", { required: "This field is required" })}
-        />
-      </FormRow>
-
-      <FormRow label="Maximum capacity" error={errors?.maxCapacity?.message}>
-        <Input
-          type="number"
-          id="maxCapacity"
-          disabled={isCreating}
-          {...register("maxCapacity", {
-            required: "This field is required",
-            min: {
-              value: 1,
-              message: "Capacity must be at least 1",
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow label="Regular price" error={errors?.regularPrice?.message}>
-        <Input
-          type="number"
-          id="regularPrice"
-          disabled={isCreating}
-          {...register("regularPrice", {
-            required: "This field is required",
-            min: {
-              value: 1,
-              message: "Regular price must be at least 1",
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow label="Discount" error={errors?.discount?.message}>
-        <Input
-          type="number"
-          id="discount"
-          defaultValue={0}
-          disabled={isCreating}
-          {...register("discount", {
-            required: "This field is required",
-            validate: (value) => {
-              value < getValues().regularPrice ||
-                "Discount should be less than regular price";
-              console.log(
-                "Discount:",
-                value,
-                "Regular Price:",
-                getValues().regularPrice
-              );
-            },
-          })}
-        />
-      </FormRow>
-
-      <FormRow
-        label="Description for website"
-        disabled={isCreating}
-        error={errors?.description?.message}
-      >
-        <Textarea
-          type="number"
-          id="description"
-          disabled={isCreating}
-          defaultValue=""
-          {...register("description", { required: "This field is required" })}
-        />
-      </FormRow>
-
-      <FormRow label="Cabin photo" error={errors.image?.message}>
-        <FileInput
-          id="image"
-          accept="image/*"
-          {...register("image", { required: "This field is required" })}
-        />
-      </FormRow>
-
-      <FormRow>
-        <Button variation="secondary" type="reset">
-          Cancel
-        </Button>
-        <Button disabled={isCreating}>Add cabin</Button>
-      </FormRow>
-    </Form>
-  );
-}
-
-export default CreateCabinForm;
