+++ src/features/dashboard/DurationChart.jsx
@@ -8,7 +8,7 @@ import {
   ResponsiveContainer,
   Tooltip,
 } from "recharts";
-import { useDarkMode } from "../../context/DarkModeContext";
+import { useDarkMode } from "../../context/useDarkMode";
 
 const ChartBox = styled.div`
   /* Box */
@@ -113,31 +113,6 @@ const startDataDark = [
     color: "#7e22ce",
   },
 ];
-
-// function prepareData(startData, stays) {
-//   function incArrayValue(arr, field) {
-//     return arr.map((obj) =>
-//       obj.duration === field ? { ...obj, value: obj.value + 1 } : obj
-//     );
-//   }
-
-//   const data = stays
-//     .reduce((arr, cur) => {
-//       const num = cur.numNights;
-//       if (num === 1) return incArrayValue(arr, "1 night");
-//       if (num === 2) return incArrayValue(arr, "2 nights");
-//       if (num === 3) return incArrayValue(arr, "3 nights");
-//       if ([4, 5].includes(num)) return incArrayValue(arr, "4-5 nights");
-//       if ([6, 7].includes(num)) return incArrayValue(arr, "6-7 nights");
-//       if (num >= 8 && num <= 14) return incArrayValue(arr, "8-14 nights");
-//       if (num >= 15 && num <= 21) return incArrayValue(arr, "15-21 nights");
-//       if (num >= 21) return incArrayValue(arr, "21+ nights");
-//       return arr;
-//     }, startData)
-//     .filter((obj) => obj.value > 0);
-
-//   return data;
-// }
 
 function prepareData(startData, stays) {
   function incArrayValue(arr, field) {
