+++ src/features/dashboard/SalesChart.jsx
@@ -10,7 +10,7 @@ import {
   XAxis,
   YAxis,
 } from "recharts";
-import { useDarkMode } from "../../context/DarkModeContext";
+import { useDarkMode } from "../../context/useDarkMode";
 import { eachDayOfInterval, subDays, format, isSameDay } from "date-fns";
 
 const StyledSalesChart = styled(DashboardBox)`
